+++ packages/terra-date-picker/CHANGELOG.md
@@ -3,8 +3,8 @@ Changelog
 
 Unreleased
 ----------
-### Changed
-* Fixed propType warning if selectedDate prop is an empty string.
+### Fixed
+* Fixed crash when selectedDate prop is set to a string or empty string that is not ISO compliant and unable to convert to a valid moment object.
 
 3.12.0 - (April 9, 2019)
 ------------------